requirements.txt
django==5.0.6
djangorestframework==3.15.1
psycopg2==2.9.9
drf-spectacular==0.27.2
Pillow==10.3.0